requirements.txt
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.1
scikit_learn==1.4.2
matplotlib==3.8.4
plotly==5.22.0
scipy==1.13.0
seaborn==0.13.2
ipykernel==6.29.4
ipython==8.12.3
kaleido==0.2.1
nbconvert==7.16.4